requirements.txt
alembic==1.8.1
jsonify
aniso8601==9.0.1
apispec==5.2.2
apispec-webframeworks==0.5.2
attrs==22.1.0
autoflake==1.7.7
beautifulsoup4==4.11.1
black==22.10.0
blinker==1.5
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
filelock==3.8.0
flake8==5.0.4
Flask==2.1.2
flask-apispec==0.11.4
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-WebTest==0.0.9
greenlet==1.1.3
gunicorn==20.1.0
identify==2.5.8
importlib-metadata==4.12.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.14.0
Mako==1.2.2
MarkupSafe==2.1.1
marshmallow==3.17.1
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
passlib==1.7.4
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
pre-commit==2.20.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pydantic==1.9.2
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
virtualenv==20.16.7
waitress==2.1.2
webargs==8.2.0
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==2.1.2
zipp==3.8.1
requests==2.26.0